(feat/refactor) rename socketSetupt to socketConfig. Tweak and tweak server-side socket 'join' event handling
