+++ client/lib/socketListeners.js
@@ -34,6 +34,4 @@ export function configureListeners(socket) {
     console.log('received socket event to remove rider:', data);
     dispatch(rideActions.removeRider(data.riderId));
   });
-
-  socket.emit('join', { user_id: 55 });
 };
