+++ client/lib/storeConfig.js
@@ -7,7 +7,7 @@ import { routerMiddleware } from 'react-router-redux';
 import { fromJS } from 'immutable';
 import { combineReducers } from 'redux-immutable';
 
-import { socket, socketActionMiddleware } from './socketSetup';
+import { socket, socketActionMiddleware } from './socketConfig';
 
 import routeReducer from '../reducers/route';
 import userReducer from '../reducers/user';
